Intermediate stage of editing the -hom-en.tex file  Changes to be committed: 	modified:   grisbi-manuel-home-en.tex 	modified:   grisbi-manuel-urldef-en.tex
Again there are no major errors so I can still generate a pdf manual to
proof read.  So far this chapter is reasonably correct up to section 4.4
(Menu Bar).  I also need to correct a misunderstanding in my
introductory Translators Note to Section 4.  Jean-Luc Duflot was NOT
editing this section for Grisbi 1.0, he is working on the documentation
for Grisbi 1.1.  My mistake Jean-Luc!